7-19 Standardized Configuration
Encapsulation of Configuration: By storing configuration data in a separate configuration file rather than hardcoding it in the code, you decouple the configuration from the codebase. This makes it easier to modify and manage the configurations without altering the code itself.

Store the Configuration in the yml file.
Then access the configuration data by defining two interface WxAccountConfig and AliPayAccountConfig