11-2 Carts Form validate
Config to not intercept the /carts api, however it does not work.

Postman show that need to log in first.

So we need to check if the request url is right or not.
Then make a breakpoint at the "UserLoginInterceptor" and the preHandle method.

It found the there is a /error page is intercepted.

The Variables show that requestDispatcherPath = "/error"
requestURL = "/error"
servletPath = "/error"

Therefor, add "/error" page in the InterceptorConfig at the same time.

After that, the console show a MethodArgumentNotValidException.
Write it down at the RuntimeExceptionHandler